LPS-86406 Move to where it is used.
